Move GenericSchemaManager, SystemSchemaManager to schema/manager package (#30389)
